requirements.txt
keras==2.4.3
numpy==1.19.5
pandas==1.0.0
requests==2.31.0
matplotlib==3.6.0
scikit-learn==1.3.2
tensorflow-gpu==2.4.0
Keras-Preprocessing==1.1.2